pyproject.toml
[project]
name = "antd-to-html-service"
version = "0.1.0"
description = "FastAPI service for converting Ant Design JSON definitions into HTML and handling form submissions."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "psycopg[binary]>=3.1",
  "psycopg_pool>=3.1",
  "python-dotenv>=1.0",
  "pydantic>=2.5"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "httpx>=0.25",
  "requests>=2.31"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
